modified function based views to class based views
