+++ store/urls.py
@@ -1,15 +1,16 @@
 from django.urls import path
-from .views import index,sign_up,login
+from views import index,Sign_up,Login
 from django.conf.urls.static import static
 from django.conf import settings
 
 
 
+
 urlpatterns = [
     path('',index,name="homepage"),
-    path('sign_up/',sign_up.as_view()),
-    path('login/',login.as_view(),name='login')
-    # path('sign_up/',sign_up)
+    path('sign_up/',Sign_up.as_view(),name='signup'),
+    path('login/',Login.as_view(),name='login'),
+
