+++ store/views/__init__.py
@@ -1,0 +1,3 @@
+from .login import Login
+from .signup import Sign_up
+from .views import index
